test: move common to workspace vitest
