+++ packages/common/package.json
@@ -82,9 +82,7 @@
   },
   "devDependencies": {
     "@types/debug": "^4.1.7",
-    "@viem/anvil": "^0.0.7",
-    "viem": "2.21.19",
-    "vitest": "0.34.6"
+    "viem": "2.21.19"
   },
   "peerDependencies": {
     "@aws-sdk/client-kms": "3.x",
