+++ test-setup/anvil.ts
@@ -1,0 +1,5 @@
+import { beforeAll, beforeEach } from "vitest";
+import { snapshotAnvilState } from "./common";
+
+beforeAll(snapshotAnvilState);
+beforeEach(snapshotAnvilState);
